# docs/swagger.yaml
definitions:
  commons.Filter:
    properties:
      field:
        type: string
      operator:
        type: string
      value:
        type: string
    type: object
  commons.Pagination2ndRequest:
    properties:
      filter:
        items:
          $ref: '#/definitions/commons.Filter'
        type: array
      order:
        additionalProperties:
          type: string
        type: object
      page:
        type: integer
      search: {}
      size:
        type: integer
    type: object
...
  models.Role:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      entity_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      is_locked:
        type: boolean
      name:
        type: string
      owner_id:
        type: string
      submitted_at:
        type: string
      submitted_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /role/deleteById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Role
  /role/getAll:
    get:
      consumes:
      - application/json
      parameters:
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get All
      tags:
      - Role
  /role/getById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Role
  /role/getPagination:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commons.Pagination2ndRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Role
  /role/getViewById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Role
...
  /role/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Role
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"